pyproject.toml
[tool.poetry]
name = "giza-cli"
version = "0.0.1"
description = "CLI for interacting with Giza Platform"
authors = ["Gonzalo Mellizo-Soto <gonzalo@gizatech.xyz>"]
readme = "README.md"
packages = [{include = "giza"}]

[tool.poetry.scripts]
giza = "giza.cli:entrypoint"

[tool.poetry.dependencies]
python = "^3.11"
requests = ">=2.30.1,<3.0.0"
tenacity = ">=8.2.2,<=9.0.0"
click = ">=8.1.0,<9.0.0"
rich = "12.6.0"
typer = {version = "0.7.0", extras = ["all"]}
pydantic = ">=1.10.8,<2.0.0"
cryptography = ">=41.0.0,<42.0.0"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
email-validator = "^2.0.0.post2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
black = ">=21,<23.3"
isort = {extras = ["pyproject"], version = "^5.12.0"}
pytest = "^7.3.2"
pytest-cov = "^4.1.0"
mypy = "^1.4.0"
ruff = "0.0.276"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[tool.ruff]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
